% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGIC.R
\name{MAGIC}
\alias{MAGIC}
\title{Main function for MAGIC}
\usage{
MAGIC(
  beta.exposure,
  beta.mediator,
  beta.outcome,
  se.exposure,
  se.mediator,
  se.outcome,
  Conf.level = 0.95,
  pval.select = c(5e-05, 5e-05),
  eta = c(0.5, 0.5),
  seed = 0
)
}
\arguments{
\item{beta.exposure}{SNP effect size's vector of the exposure vairable (GWASI)}

\item{beta.mediator}{SNP effect size's vector of the mediator vairable (GWASIII)}

\item{beta.outcome}{SNP effect size's vector of the outcome vairable (GWASII)}

\item{se.exposure}{SNP effect size's standard errors of \code{beta.exposure}}

\item{se.mediator}{SNP effect size's standard errors of \code{beta.mediator}}

\item{se.outcome}{SNP effect size's standard errors of \code{beta.outcome}}

\item{Conf.level}{Confidence level. Default is 0.95.}

\item{pval.select}{A vector of specified pre-screening threshold in the ordering of (exposure, mediator). Default is c(5e-5, 5e-5). (corresponding lambda is 4.06)}

\item{eta}{A vector of rerandomized scale in the ordering of  (exposure, mediator). Default is c(0.5,0.5).}

\item{seed}{The value of random seed. Default is 0.}
}
\value{
A list
\describe{
\item{theta.hat}{Estimated direct effect from exposure to outcome variable}
\item{tauy.hat}{Estimated direct effect from mediator to outcome variable}
\item{taux.hat}{Estimated indirect effect from exposure to mediator variable}
\item{tau.hat}{Estimated mediation effect}
\item{tau_total.hat}{Estimated total effect}
\item{theta.se}{Standard error of \code{theta.hat}}
\item{tauy.se}{Standard error of \code{tauy.hat}}
\item{taux.se}{Standard error of \code{taux.hat}}
\item{tau.se}{Standard error of \code{tau.hat}}
\item{tau_total.se}{Standard error of \code{tau_total.hat}}
\item{n.IV.exp}{Number of IVs used in exposure dataset}
\item{n.IV.med}{Number of IVs used in mediator dataset}
\item{Conf.Interval}{Confidence interval given \code{Conf.level}}
\item{IV.exp}{The index of IVs selected in Sx}
\item{IV.med}{The index of IVs selected in Sm}
}
}
\description{
Main function for MAGIC
}
\references{
Rita Qiuran Lyu, Chong Wu, Xinwei Ma, Jingshen Wang (2023). Mediation Analysis with Mendelian Randomization and Efficient Multiple GWAS Integration. \url{https://arxiv.org/abs/2312.10563}.
}
