% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIVW.R
\name{pre_screening}
\alias{pre_screening}
\title{Supplementary function for RIVW}
\usage{
pre_screening(
  gamma1.exp,
  se1.exp,
  etamean = 0.5,
  pthr = 5e-05,
  seed = 0,
  smoothing = FALSE
)
}
\arguments{
\item{gamma1.exp}{SNP effect size's vector of the exposure vairable}

\item{se1.exp}{SNP effect size's standard errors of \code{beta.exposure}}

\item{etamean}{rerandomized scale of exposure variable. Default is 0.5.}

\item{pthr}{The specified pre-screening threshold. Default is 5e-5. (corresponding lambda is 4.06)}

\item{seed}{A random seed. Default is 0.}

\item{smoothing}{Whether to use smoothing to decrease variance . Default is FALSE.}
}
\value{
A list
\describe{
\item{filter1}{Indexs of selected relevant IVS.}
\item{gamma_exp1}{Effect size in GWAS (I) after Rao-Blackwellization to eliminate the winner's curse}
\item{se1}{Standard errors in GWAS (I) after Rao-Blackwellization to eliminate the winner's curse}
\item{weights}{The weights for each SNP. If smoothing is False, weights are the same for each SNP.}
}
}
\description{
Supplementary function for RIVW
}
\references{
Xinwei Ma, Jingshen Wang, Chong Wu. (2023). Breaking the Winner’s Curse in Mendelian Randomization:Rerandomized Inverse Variance Weighted Estimator \url{https://projecteuclid.org/journals/annals-of-statistics/volume-51/issue-1/Breaking-the-winners-curse-in-Mendelian-randomization--Rerandomized-inverse/10.1214/22-AOS2247.full}.
}
