% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGIC.R
\name{pre_selection}
\alias{pre_selection}
\title{Supplementary function for MAGIC}
\usage{
pre_selection(
  gamma1.exp,
  se1.exp,
  gamma2.exp,
  se2.exp,
  etamean1 = 0.5,
  etamean2 = 0.5,
  pthr = c(5e-05, 5e-05),
  seed = 0
)
}
\arguments{
\item{gamma1.exp}{SNP effect size's vector of the exposure vairable (GWASI)}

\item{se1.exp}{SNP effect size's standard errors of \code{beta.exposure}}

\item{gamma2.exp}{SNP effect size's vector of the mediator vairable (GWASIII)}

\item{se2.exp}{SNP effect size's standard errors of \code{beta.mediator}}

\item{etamean1}{rerandomized scale of exposure variable. Default is 0.5.}

\item{etamean2}{rerandomized scale of mediator variable. Default is 0.5.}

\item{pthr}{A vector of specified pre-screening threshold in the ordering of (exposure, mediator). Default is c(5e-5, 5e-5). (corresponding lambda is 4.06)}

\item{seed}{a random seed. Default is 0.}
}
\value{
A list
\describe{
\item{filter1}{Indexs of selected relevant IVs in Sx}
\item{filter2}{Indexs of selected relevant IVs in Sm}
\item{gamma_exp1}{Effect size in GWAS (I) after Rao-Blackwellization to eliminate the winner's curse}
\item{se1}{Standard errors in GWAS (I) after Rao-Blackwellization to eliminate the winner's curse}
\item{gamma_exp2}{Effect size in GWAS (III) after Rao-Blackwellization to eliminate the winner's curse}
\item{se2}{Standard errors in GWAS (III) after Rao-Blackwellization to eliminate the winner's curse}
\item{gamma_exp1.carve}{Effect size in GWAS (I) after Rao-Blackwellization to eliminate the loser's curse}
\item{gamma_exp2.carve}{Effect size in GWAS (III) after Rao-Blackwellization to eliminate the loser's curse}
}
}
\description{
Supplementary function for MAGIC
}
\references{
Rita Qiuran Lyu, Chong Wu, Xinwei Ma, Jingshen Wang (2023). Mediation Analysis with Mendelian Randomization and Efficient Multiple GWAS Integration. \url{https://arxiv.org/abs/2312.10563}.
}
